\name{LIDAR}
\docType{data}

\alias{LIDAR}


\title{Canopy LIDAR data}

\description{ 
This is a small portion of Light Detection and Ranging (LIDAR) data
taken over a forested landscape in Wisconsin, USA.
}

\usage{
data(LIDAR)
}

\format{
 A data frame containing 10123 rows and 3 columns corresponding to
 longitude, latitude, and elevation.
}

\source{
  Data provided by:
  Dr. Paul V. Bolstad,
  Department of Forest Resources,
  University of Minnesota,
  \email{pbolstad@umn.edu}
}


\keyword{datasets}

